+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - CLI option to load an audit library: `-L` or `--audit-library`
+- Implemented `SLL_ABI_AUDIT_CALL`, `SLL_ABI_AUDIT_DEINIT`, `SLL_ABI_AUDIT_INIT`, `SLL_ABI_DEINIT`, `SLL_ABI_INIT` and `SLL_ABI_NAME`
 - Implemented `statistics$correlation` ([#260])
 - Object creation code to build an array of string pointers: `x`